fix: 'sdk-react-ui' MetaMaskModal is not showing balance
