+++ package.json
@@ -23,6 +23,8 @@
   "scripts": {
     "build": "cd packages/sdk-socket-server && yarn install && cd ../.. && yarn workspaces foreach --verbose run build:pre-tsc && yarn workspaces foreach --verbose --topological --parallel --no-private run build && yarn workspaces foreach --verbose run build:post-tsc ",
     "build:clean": "yarn clean && yarn build",
+    "light-build:": "cd packages/sdk-socket-server && yarn install && cd ../.. && yarn workspaces foreach --exclude @metamask/sdk-ui --verbose run build:pre-tsc && yarn workspaces foreach --exclude @metamask/sdk-ui --verbose --topological --parallel --no-private run build && yarn workspaces foreach --exclude @metamask/sdk-ui --verbose run build:post-tsc",
+    "light-build:clean": "yarn clean && yarn build:light",
     "build:tsc": "tsc --build --force tsconfig.json",
     "clean": "yarn workspaces foreach --no-private run clean",
     "postinstall": "patch-package",
