+++ packages/examples/react-metamask-button/package.json
@@ -28,7 +28,7 @@
     ]
   },
   "dependencies": {
-    "@metamask/sdk-react-ui": "^0.12.2",
+    "@metamask/sdk-react-ui": "^0.14.2",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
